requirements.txt
gradio
PyPDF2
requests
beautifulsoup4
lxml
python-dotenv
